+++ packages/actions/src/schema.ts
@@ -2,7 +2,6 @@ export type Schema = {
   forceAttribute: string;
   focusGroupAttribute: string;
   focusDirectionAttribute: string;
-  permanentAttribute: string;
   hiddenClassName: string;
 };
 
@@ -10,6 +9,5 @@ export const defaultSchema: Schema = {
   forceAttribute: 'data-turbo-force',
   focusGroupAttribute: 'data-turbo-focus-group',
   focusDirectionAttribute: 'data-turbo-focus-direction',
-  permanentAttribute: 'data-turbo-permanent',
   hiddenClassName: 'hidden',
 };