+++ packages/turbo-stream/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @coldwired/turbo-stream
 
+## 0.11.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @coldwired/actions@0.11.0
+
 ## 0.10.0
 
 ### Patch Changes